NEWTAG-518 Model already integrated, so this is just deprecation
