+++ code/src/java/plugin/exporttokens/deprecated/SizeToken.java
@@ -1,5 +1,5 @@
 /*
- * SizeLongToken.java
+ * SizeToken.java
  * Copyright 2003 (C) Devon Jones <soulcatcher@evilsoft.org>
  *
  * This library is free software; you can redistribute it and/or
@@ -23,17 +23,16 @@
  * Last Edited: $Date$
  *
  */
-package plugin.exporttokens;
+package plugin.exporttokens.deprecated;
 
-import pcgen.core.SizeAdjustment;
 import pcgen.core.display.CharacterDisplay;
 import pcgen.io.ExportHandler;
 import pcgen.io.exporttoken.AbstractExportToken;
 
 /**
- * SIZELONG for export
+ * SIZE token for export
  */
-public class SizeLongToken extends AbstractExportToken
+public class SizeToken extends AbstractExportToken
 {
 	/**
 	 * @see pcgen.io.exporttoken.Token#getTokenName()
@@ -41,10 +40,9 @@ public class SizeLongToken extends AbstractExportToken
 	@Override
 	public String getTokenName()
 	{
-		return "SIZELONG";
+		return "SIZE";
 	}
 
-	//TODO: this really should be in the Size token as SIZE.LONG
 	/**
 	 * @see pcgen.io.exporttoken.Token#getToken(java.lang.String, pcgen.core.PlayerCharacter, pcgen.io.ExportHandler)
 	 */
@@ -52,11 +50,6 @@ public class SizeLongToken extends AbstractExportToken
 	public String getToken(String tokenSource, CharacterDisplay display,
 		ExportHandler eh)
 	{
-		final SizeAdjustment sadj = display.getSizeAdjustment();
-		if (sadj != null)
-		{
-			return sadj.getDisplayName();
-		}
-		return "";
+		return display.getSize();
 	}
 }
